+++ tests/test_pg8000.py
@@ -1,14 +1,17 @@
 from datetime import date
 
 import aiosql
-import pg8000
+import pg8000 as db
 import pytest
 import run_tests as t
 
 
+DRIVER = "pg8000"
+
+
 @pytest.fixture()
 def queries():
-    return t.queries("pg8000")
+    return t.queries(DRIVER)
 
 
 @pytest.fixture()
@@ -17,13 +20,13 @@ def conn(pg_params):
     del pg_params["dbname"]
     pg_params["database"] = dbname
     t.log.debug(f"params: {pg_params}")
-    with pg8000.connect(**pg_params) as conn:
+    with db.connect(**pg_params) as conn:
         yield conn
 
 
 @pytest.mark.skip("row factory is not supported")
 def test_record_query(pg_params, queries):
-    with pg8000.connect(**pg_params, row_factory=dict_row) as conn:
+    with db.connect(**pg_params, row_factory=dict_row) as conn:
         t.run_record_query(conn, queries)
 
 
@@ -33,7 +36,7 @@ def test_parameterized_query(conn, queries):
 
 @pytest.mark.skip("row factory is not supported")
 def test_parameterized_record_query(pg_params, queries):
-    with pg8000.connect(**pg_params, row_factory=dict_row) as conn:
+    with db.connect(**pg_params, row_factory=dict_row) as conn:
         t.run_parameterized_record_query(conn, queries, "pg", date)
 
 
@@ -60,7 +63,7 @@ def test_modulo(conn, queries):
 
 
 def test_insert_returning(conn, queries):
-    t.run_insert_returning(conn, queries, "pg8000", date)
+    t.run_insert_returning(conn, queries, DRIVER, date)
 
 
 def test_delete(conn, queries):
